Add files via upload
This is a minor project on Exploratory Data Analysis. The dataset for the project is a record of some students' academic performances in Maths, Reading, and Writing. The objective of the EDA was to understand how the student's performance (test scores) is affected by the other variables (Gender, Ethnicity, Parental level of education, Lunch, Test preparation course).